Checkbox Component with Tailwind (#654)
* Created Checkbox Component with Tailwind

* Deleted .scss file and refactored away old Checkbox component

* Updated Checkbox

* Installed tailwind-merge and implemented cn function

* Updated Checkbox with twmerge

* Update frontend/src/tw-components/Checkbox.tsx


* Update frontend/src/tw-components/Checkbox.tsx


* Update frontend/tests/components/Checkbox.test.tsx


* Feat/qualifier pages (#668)

* update qualifier page 1

- aligned qualifier nav
- use grid system instead of flex box for items

* Refactor: update button styles for primary-dark theme

* Feat: add ProgressIndicator component and integrate it into QualifierPage2

* Refactor: update border style in QualifierNav component

* Refactor: dynamic grid layout in QualifierPage1 for responsive design

* Feat: enhance RadioButtonForm to support skill selection callback

* Feat: update RadioButtonForm dynamically render skill & experience levels

* Add selectCOP and experience levels props to qualifiers

* Feat: use context in QualiferPage2 to keep track of chosen experience level

* Feat: integrate qualifiers context to manage selected COP in QualifierPage1

* Feat: refactor QualifierPage2 layout and improve skill selection handling

* Refactor:

- renamed experienceLevels to skills_matrix
- refactored to use tw-components/Buttons.tsx & added second style
- Ran npm run lint in frontend directory

* update tech stack

* feat: use new tw-button variant, fix merge

---------


* Fixed className syntax

---------
