--- frontend/src/components/Inputs/_Checkbox.scss
@@ -1,51 +1,0 @@
-// External Imports
-@use "sass:color";
-
-// Internal Imports
-@use "../Basics" as *;
-
-/***************************
-*** Classes for checkbox ***
-***************************/
-.checkbox {
-  display: block;
-}
-
-.checkbox-icon {
-  align-items: center;
-  border-radius: 50%;
-  display: inline-flex;
-  height: 40px;
-  justify-content: center;
-  width: 40px;
-}
-
-:not(.checkbox-disabled) > .checkbox-icon {
-  &:hover,
-  &:focus,
-  .checkbox-input:focus-visible + & {
-    background-color: color.change($color-blue-dark, $alpha: 0.16);
-  }
-}
-
-.checkbox-icon-enter,
-.checkbox-icon-exit {
-  background-color: color.change($color-blue-dark, $alpha: 0.32) !important;
-}
-
-.checkbox-icon-enter-active,
-.checkbox-icon-exit-active {
-  background-color: color.change($color-blue-dark, $alpha: 0.16) !important;
-  transition: background-color 300ms;
-}
-
-.checkbox-label {
-  align-items: center;
-  display: flex;
-  user-select: none;
-  width: fit-content;
-
-  &:not(.checkbox-disabled):hover {
-    cursor: pointer;
-  }
-}
