+++ frontend/src/components/Inputs/_index.scss
@@ -1,5 +1,4 @@
 @forward "Calendar";
-@forward "Checkbox";
 @forward "Chip";
 @forward "Dropdown";
 @forward "Textfield";
