+++ frontend/src/components/components.tsx
@@ -2,7 +2,6 @@ import { Button } from "./Buttons/Button";
 import { IconButton } from "./Buttons/IconButton";
 import { ScrollCarousel } from "./Carousel/ScrollCarousel";
 import { Calendar } from "./Inputs/Calendar";
-import { Checkbox } from "./Inputs/Checkbox";
 import { Chip } from "./Inputs/Chip";
 import { Dropdown, DropdownOption } from "./Inputs/Dropdown";
 import { TextField } from "./Inputs/Textfield";
@@ -18,7 +17,6 @@ export {
   // Dialog
   // Inputs
   Calendar,
-  Checkbox,
   Chip,
   Dropdown,
   DropdownOption,
